Refactor `Playlists` container to `PlaylistList`
`Playlists` did not really need a container as it only passed the data
onto `PlaylistList`.
